+++ components/trade/GraphData.jsx
